v14

+++ commands/cmds/clear.js
@@ -8,7 +8,7 @@ module.exports = {
     description: "this description is weird",
     callback: async(message, arguments, text) => {
 
-        let clearmsg = +arguments[0]
+        let clearmsg = Math.round(+arguments[0])
         var iconurl = message.guild.iconURL({ dynamic: true })
         const modlog = '822575095721099304'
         const Discord = require('discord.js')
@@ -24,18 +24,19 @@ module.exports = {
             if (clearmsg === 1) {
                 await message.channel.bulkDelete(clearmsg + 1)
                 message.channel.send('Es wurde `eine` Nachricht gelöscht. \nAusgeführt durch: ' + message.author.tag)
-            } else if (clearmsg >= 2) {
+            } else if (clearmsg >= 2 && clearmsg <= 100) {
+                if (clearmsg == 100) clearmsg = 99
                 await message.channel.bulkDelete(clearmsg + 1)
                 message.channel.send('Es wurden `' + clearmsg + '` Nachrichten gelöscht. \nAusgeführt durch: ' + message.author.tag)
-            } else message.channel.send("Du kannste keine " + clearmsg + " Nachrichten löschen, " + message.author.username)
+            } else message.channel.send("Du kannste keine `" + clearmsg + "` Nachrichten löschen, " + message.author.username + ".\nDas Limit liegt bei 100 Nachrichten.")
 
         } catch (error) {
-            const embedfail = new Discord.MessageEmbed()
+            const embedfail = new Discord.EmbedBuilder()
                 .setTitle('Es gab einen Fehler beim Löschen von Nachrichten')
                 .setThumbnail('https://mir-s3-cdn-cf.behance.net/project_modules/max_1200/ab0c1e57515093.59d8c6eb16d19.gif')
                 .setDescription('Fehler: `' + error + '`')
-                .addField('Von: ' + message.author.tag, 'in <#' + message.channel.id + '>')
-                .setFooter('thevalleyy-NetWork', iconurl)
+                .addFields([{name: 'Von: ' + message.author.tag, value: 'in <#' + message.channel.id + '>'}])
+                .setFooter({ text: 'thevalleyy-NetWork', iconURL: iconurl})
                 .setTimestamp()
                 .setColor('fc036b')
             message.client.channels.cache.get(modlog).send({ embeds: [embedfail] });
